requirements.txt
anyio==4.8.0
APScheduler==3.11.0
asgiref==3.8.1
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==4.2
et_xmlfile==2.0.0
filelock==3.18.0
Flask==3.1.0
fsspec==2025.5.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.33.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
openpyxl==3.1.5
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pyTelegramBotAPI==4.26.0
python-decouple==3.8
python-dotenv==1.0.1
python-telegram-bot==21.10
pytube==15.0.0
pytz==2025.1
PyYAML==6.0.2
qrcode==8.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==4.1.0
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
telebot==0.0.5
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
urllib3==2.3.0
waitress==3.0.2
Werkzeug==3.1.3
aiogram>=3.6
python-dateutil>=2.9